Allow locking to window size

+++ M64RPFW.Views.Avalonia/Helpers/ServiceConversions.cs
@@ -7,6 +7,6 @@ public static class ServiceConversions
 {
     public static WindowPoint ToWindowPoint(this Point point)
     {
-        return new WindowPoint(point.X, point.Y);
+        return new WindowPoint((int) point.X, (int) point.Y);
     }
 }
